requirements.txt
fastapi~=0.123.10
uvicorn
sqlalchemy~=2.0.44
psycopg2-binary
python-dotenv
pydantic~=2.12.5
bcrypt~=4.3.0
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.5.0
pydantic-settings~=2.12.0
bcrypt~=4.3.0